BusStopList now use ABC ActionBar.
The problem is ABC ActionBar with ListFragment does not work on older
Android but if native ActionBar is used the search view does not work
(as search.xml use support namespace and refer to support library's
SearchView). I think it is time to make tradeoff and trade older Android
support for the greater good of the newer one.

At this time Android 2.2-2.3.7 are 8.57% of the users. with the majority
on Android 4.1, 28.31%.