Initial work on BusMapFragment
The old BusMapFragment is moved to MapFragment. The BusStopInfoFragment is still under
decision whether what to rewrite into activity or other choices. The other places
that use BusMapController is now rewritten to use BusMapFragment.

BusMapController is deprecated.

The goal of BusMapFragment is to ensure cleaner code and easier to use API
(as seen in RouteMapFragment that only one method remains as the activity lifecycle
is now managed by the fragment system instead of manually)

Also cleaned up several warnings